Make sc_apparmor public
