+++ src/apparmor-support.c
@@ -21,7 +21,6 @@
 
 #include "apparmor-support.h"
 
-#include <stdbool.h>
 #include <string.h>
 #ifdef HAVE_APPARMOR
 #include <sys/apparmor.h>
@@ -34,26 +33,6 @@
 // changed without breaking apparmor functionality.
 #define SC_AA_ENFORCE_STR "enforce"
 #define SC_AA_COMPLAIN_STR "complain"
-
-enum sc_apparmor_mode {
-	// The enforcement mode was not recognized.
-	SC_AA_INVALID = -1,
-	// The enforcement mode is not applicable because apparmor is disabled.
-	SC_AA_NOT_APPLICABLE = 0,
-	// The enforcement mode is "enforcing"
-	SC_AA_ENFORCE = 1,
-	// The enforcement mode is "complain"
-	SC_AA_COMPLAIN,
-};
-
-struct sc_apparmor {
-	// The mode of enforcement. In addition to the two apparmor defined modes
-	// can be also SC_AA_INVALID (unknown mode reported by apparmor) and
-	// SC_AA_NOT_APPLICABLE (when we're not linked with apparmor).
-	enum sc_apparmor_mode mode;
-	// Flag indicating that the current process is confined.
-	bool is_confined;
-};
 
 void sc_init_apparmor_support(struct sc_apparmor *apparmor)
 {
