+++ src/apparmor-support.h
@@ -18,10 +18,30 @@
 #ifndef SNAP_CONFINE_APPARMOR_SUPPORT_H
 #define SNAP_CONFINE_APPARMOR_SUPPORT_H
 
+#include <stdbool.h>
+
 /**
  * Data required to manage apparmor wrapper. 
  */
-struct sc_apparmor;
+enum sc_apparmor_mode {
+	// The enforcement mode was not recognized.
+	SC_AA_INVALID = -1,
+	// The enforcement mode is not applicable because apparmor is disabled.
+	SC_AA_NOT_APPLICABLE = 0,
+	// The enforcement mode is "enforcing"
+	SC_AA_ENFORCE = 1,
+	// The enforcement mode is "complain"
+	SC_AA_COMPLAIN,
+};
+
+struct sc_apparmor {
+	// The mode of enforcement. In addition to the two apparmor defined modes
+	// can be also SC_AA_INVALID (unknown mode reported by apparmor) and
+	// SC_AA_NOT_APPLICABLE (when we're not linked with apparmor).
+	enum sc_apparmor_mode mode;
+	// Flag indicating that the current process is confined.
+	bool is_confined;
+};
 
 /**
  * Initialize apparmor support.
